# implemented efficient readline for ByteBuffer