--- a/integration-tests/test_s3_readline.py
+++ b/integration-tests/test_s3_readline.py
@@ -1,4 +1,3 @@
-import sys
 from smart_open import open