#### Motivation

The previous implementation searched for the terminator in the buffer twice. This was unnecessary.

The new implementaiton performs the search only once.

- Fixes #371 

#### Checklist

Before you create the PR, please make sure you have:

- [x] Picked a concise, informative and complete title
- [x] Clearly explained the motivation behind the PR
- [x] Linked to any existing issues that your PR will be solving
- [x] Included tests for any new functionality
- [x] Checked that all unit tests pass