@mpenkov 2020-03-11
Please check out the benchmarks above. With 5 runs to read 1 million lines, the minimum time hasn't decreased all that much (from 38 to 36s), but the max time changed significantly (from 58 to 40s).